primitives::block_header

Struct BlockHeader

pub struct BlockHeader {
    pub pow_hash: Option<H256>,
    pub approximated_rlp_size: usize,
    /* private fields */
}

A block header.

Fields
pow_hash: Option<H256>

POW quality of the block

approximated_rlp_size: usize

Approximated rlp size of the block header

Implementations

impl BlockHeader

pub fn approximated_rlp_size(&self) -> usize

Approximated rlp size of the block header.

pub fn parent_hash(&self) -> &H256

Get the parent_hash field of the header.

pub fn height(&self) -> u64

Get the block height

pub fn timestamp(&self) -> u64

Get the timestamp field of the header.

pub fn author(&self) -> &Address

Get the author field of the header.

pub fn transactions_root(&self) -> &H256

Get the transactions root field of the header.

pub fn deferred_state_root(&self) -> &H256

Get the deferred state root field of the header.

pub fn deferred_receipts_root(&self) -> &H256

Get the deferred block receipts root field of the header.

pub fn deferred_logs_bloom_hash(&self) -> &H256

Get the deferred block logs bloom hash field of the header.

pub fn blame(&self) -> u32

Get the blame field of the header

pub fn difficulty(&self) -> &U256

Get the difficulty field of the header.

pub fn adaptive(&self) -> bool

Get the adaptive field of the header

pub fn gas_limit(&self) -> &U256

Get the gas limit field of the header.

pub fn core_space_gas_limit(&self) -> U256

pub fn espace_gas_limit(&self, can_pack: bool) -> U256

pub fn referee_hashes(&self) -> &Vec<H256>

Get the referee hashes field of the header.

pub fn custom(&self) -> &Vec<Bytes>

Get the custom data field of the header.

pub fn nonce(&self) -> U256

Get the nonce field of the header.

pub fn pos_reference(&self) -> &Option<PosBlockId>

Get the PoS reference.

pub fn base_price(&self) -> Option<SpaceMap<U256>>

pub fn space_base_price(&self, space: Space) -> Option<U256>

pub fn set_nonce(&mut self, nonce: U256)

Set the nonce field of the header.

pub fn set_timestamp(&mut self, timestamp: u64)

Set the timestamp filed of the header.

pub fn set_custom(&mut self, custom: Vec<Bytes>)

Set the custom filed of the header.

pub fn compute_hash(&mut self) -> H256

Compute the hash of the block.

pub fn hash(&self) -> H256

Get the hash of the block.

pub fn problem_hash(&self) -> H256

Get the hash of PoW problem.

pub fn rlp_without_nonce(&self) -> Bytes

Get the RLP representation of this header(except nonce).

pub fn rlp(&self) -> Bytes

Get the RLP representation of this header.

pub fn stream_rlp_with_pow_hash(&self, stream: &mut RlpStream)

Place this header and its pow_hash into an RLP stream stream.

pub fn decode_with_pow_hash(bytes: &[u8]) -> Result<Self, DecoderError>

pub fn size(&self) -> usize