primitives

Module compare

Functions

and_vec_bit_multithreads

and_vec_multithreads

eq_many

Returns vector of 1 if lhs[i] == rhs[i] and 0 otherwise. Checks if shared values in lhs are equal to corresponding shared values in rhs. The result is a vector of shared values that have value 1 if the corresponding lhs value is equal to the corresponding rhs value and 0 otherwise.

eq_many_binary

eq_many_binary_multithreads

eq_many_multithreads

eq_public_many

eq_public_many_binary

eq_public_many_binary_multithreads

eq_public_many_multithreads

ge_many

Returns vector of 1 if lhs[i] >= rhs[i] and 0 otherwise. Checks if shared values in lhs are greater than or equal to corresponding shared values in rhs. The result is a vector of shared values that have value 1 if the corresponding lhs value is greater than or equal to the corresponding rhs value and 0 otherwise.

ge_many_binary

ge_many_binary_multithreads

ge_many_multithreads

ge_public_many

ge_public_many_binary

ge_public_many_binary_multithreads

ge_public_many_multithreads

gt_many

Returns vector of 1 if lhs[i] > rhs[i] and 0 otherwise. Checks if shared values in lhs are greater than corresponding shared values in rhs. The result is a vector of shared values that have value 1 if the corresponding lhs value is greater than the corresponding rhs value and 0 otherwise.

gt_many_binary

gt_many_binary_multithreads

gt_many_multithreads

gt_public_many

gt_public_many_binary

gt_public_many_binary_multithreads

gt_public_many_multithreads

is_zero_many

Batched version of is_zero that checks multiple values at once. This is more communication efficient as it batches the AND operations.

is_zero_many_multithreads

Batched version of is_zero that checks multiple values at once. This is more communication efficient as it batches the AND operations.

le_many

Returns vector of 1 if lhs[i] <= rhs[i] and 0 otherwise. Checks if shared values in lhs are less than or equal to corresponding shared values in rhs. The result is a vector of shared values that have value 1 if the corresponding lhs value is less than or equal to the corresponding rhs value and 0 otherwise.

le_many_binary

le_many_binary_multithreads

le_many_multithreads

le_public_many

le_public_many_binary

le_public_many_binary_multithreads

le_public_many_multithreads

lt_many

Returns 1 if lhs < rhs and 0 otherwise. Checks if one shared value is less than another shared value. The result is a shared value that has value 1 if the first shared value is less than the second shared value and 0 otherwise.

lt_many_binary

lt_many_binary_multithreads

lt_many_multithreads

lt_public_many

lt_public_many_binary

lt_public_many_binary_multithreads

lt_public_many_multithreads

neq_many

Returns vector of 1 if lhs[i] != rhs[i] and 0 otherwise. Checks if shared values in lhs are not equal to corresponding shared values in rhs. The result is a vector of shared values that have value 1 if the corresponding lhs value is not equal to the corresponding rhs value and 0 otherwise.

neq_many_binary

neq_many_binary_multithreads

neq_many_multithreads

neq_public_many

neq_public_many_binary

neq_public_many_binary_multithreads

neq_public_many_multithreads

or_vec_multithreads

unsigned_ge_const_lhs_many

unsigned_ge_const_lhs_many_binary_multithreads

unsigned_ge_const_lhs_many_multithreads

unsigned_ge_const_rhs_many_binary_multithreads

unsigned_ge_const_rhs_many_multithreads

unsigned_ge_many_binary_multithreads

unsigned_ge_many_multithreads