primitives::compare

Function neq_many

pub fn neq_many<T: IntRing2k, N: Network>(
    a: &[Rep3RingShare<T>],
    b: &[Rep3RingShare<T>],
    net: &N,
    state: &mut Rep3State,
) -> Result<Vec<Rep3RingShare<Bit>>>
where
    Standard: Distribution<T>,

Returns vector of 1 if lhs[i] != rhs[i] and 0 otherwise. Checks if shared values in lhs are not equal to corresponding shared values in rhs. The result is a vector of shared values that have value 1 if the corresponding lhs value is not equal to the corresponding rhs value and 0 otherwise.