Struct primitives::config::Config

pub struct Config {

    pub max_channels: u32,
    pub channels_find_limit: u32,
    pub campaigns_find_limit: u32,
    pub spendable_find_limit: u32,
    pub wait_time: Duration,
    pub msgs_find_limit: u32,
    pub analytics_find_limit: u32,
    pub analytics_maxtime: Duration,
    pub heartbeat_time: Duration,
    pub health_threshold_promilles: u32,
    pub health_unsignable_promilles: u32,
    pub propagation_timeout: Duration,
    pub fetch_timeout: Duration,
    pub all_campaigns_timeout: Duration,
    pub channel_tick_timeout: Duration,
    pub ip_rate_limit: RateLimit,
    pub creators_whitelist: Vec<Address>,
    pub validators_whitelist: Vec<ValidatorId>,
    pub admins: Vec<Address>,
    pub chains: HashMap<String, ChainInfo>,
    pub platform: PlatformConfig,
    pub limits: Limits,

}

Examples:

use primitives::{config::GANACHE_CONFIG, Config};
use serde_json::{from_value, json};

fn main() {
    let json = json!({
      "max_channels": 512,
      "channels_find_limit": 200,
      "campaigns_find_limit": 200,
      "spendable_find_limit": 200,
      "wait_time": 500,
      "msgs_find_limit": 10,
      "analytics_find_limit": 5000,
      "analytics_maxtime": 20000,
      "heartbeat_time": 30000,
      "health_threshold_promilles": 950,
      "health_unsignable_promilles": 750,
      "propagation_timeout": 2000,
      "fetch_timeout": 5000,
      "all_campaigns_timeout": 5000,
      "channel_tick_timeout": 8000,
      "ip_rate_limit": {
        "type": "ip",
        "timeframe": 1200000
      },
      "creators_whitelist": [
        "0xaCBaDA2d5830d1875ae3D2de207A1363B316Df2F",
        "0xDd589B43793934EF6Ad266067A0d1D4896b0dff0",
        "0x541b401362Ea1D489D322579552B099e801F3632"
      ],
      "validators_whitelist": [
        "0x80690751969B234697e9059e04ed72195c3507fa",
        "0xf3f583AEC5f7C030722Fe992A5688557e1B86ef7",
        "0x6B83e7D6B72c098d48968441e0d05658dc17Adb9"
      ],
      "admins": [
        "0x80690751969B234697e9059e04ed72195c3507fa"
      ],
      "chain": {
        "Ganache #1337": {
          "chain_id": 1337,
          "rpc": "http://localhost:1337/",
          "outpace": "0xAbc27d46a458E2e49DaBfEf45ca74dEDBAc3DD06",
          "token": {
            "Mocked TOKEN 1337": {
              "min_campaign_budget": "1000000000000000000",
              "min_validator_fee": "1000000000000",
              "precision": 18,
              "address": "0x2BCaf6968aEC8A3b5126FBfAb5Fd419da6E8AD8E"
            }
          }
        },
        "Ganache #1": {
          "chain_id": 1,
          "rpc": "http://localhost:8545/",
          "outpace": "0x26CBc2eAAe377f6Ac4b73a982CD1125eF4CEC96f",
          "token": {
            "Mocked TOKEN 1": {
              "min_campaign_budget": "1000000000000000000",
              "min_validator_fee": "1000000000000",
              "precision": 18,
              "address": "0x12a28f2bfBFfDf5842657235cC058242f40fDEa6"
            }
          }
        }
      },
      "platform": {
        "url": "https://platform.adex.network/",
        "keep_alive_interval": 1200000
      },
      "limits": {
        "units_for_slot": {
          "max_campaigns_earning_from": 25,
          "global_min_impression_price": "1000000"
        }
      }
    });
    assert_eq!(
        from_value::<Config>(json).expect("Should deserialize"),
        GANACHE_CONFIG.clone()
    );
}

Fields
max_channels: u32

The maximum number of Channels that the worker can process for one tick.

channels_find_limit: u32

The maximum number of Channels per page returned by Sentry’s GET /v5/channel/list route.

Also see: ChannelListResponse

campaigns_find_limit: u32

The maximum number of Campaigns per page returned by Sentry’s GET /v5/campaign/list route.

Also see: CampaignListResponse

spendable_find_limit: u32

The maximum number of Spenders per page returned by Sentry’s GET /v5/channel/0xXXX.../spender/all route.

Also see: AllSpendersResponse

wait_time: Duration

The Validator Worker tick time.

The Channels’ tick and the wait time should both finish before running a new tick in the Validator Worker.

In milliseconds

msgs_find_limit: u32

The maximum allowed limit of ValidatorMessages per page returned by Sentry’s GET /v5/channel/0xXXX.../validator-messages route.

Request query also has a limit parameter, which can be used to return <= msgs_find_limit messages in the request.

Also see: ValidatorMessagesListResponse, ValidatorMessagesListQuery

analytics_find_limit: u32

The maximum allowed limit of FetchedAnalyticss per page returned by Sentry’s GET /v5/analytics routes:

GET /v5/analytics

GET /v5/analytics/for-publisher

GET /v5/analytics/for-advertiser

GET /v5/analytics/for-admin

Request query also has a limit parameter, which can be used to return <= analytics_find_limit messages in the request.

Also see: AnalyticsQuery

analytics_maxtime: Duration

A timeout to be used when collecting the Analytics for a requests:

GET /v5/analytics

GET /v5/analytics/for-publisher

GET /v5/analytics/for-advertiser

GET /v5/analytics/for-admin

In milliseconds

heartbeat_time: Duration

The amount of time that should have passed before sending a new heartbeat.

In milliseconds

health_threshold_promilles: u32

The pro miles below which the ApproveState becomes unhealthy in the Channel’s Follower.

Also see: ApproveState.is_healthy

In pro milles (<= 1000)

health_unsignable_promilles: u32

The pro milles below which the ApproveState will not be triggered and instead a RejectState will be propagated by the Channel’s Follower.

In pro milles (<= 1000)

propagation_timeout: Duration

Sets the timeout for propagating a Validator message (MessageTypes) to a validator.

In milliseconds

fetch_timeout: Duration

The Client timeout for SentryApi.

This includes all requests made to sentry except propagating messages. When propagating messages we make requests to foreign Sentry instances and we use a separate timeout - Config.propagation_timeout.

In milliseconds

all_campaigns_timeout: Duration

The Client timeout for SentryApi when collecting all channels and Validators using the /campaign/list route.

In milliseconds

channel_tick_timeout: Duration

The timeout for a single tick of a Channel in the Validator Worker. This timeout is applied to both the leader and follower ticks.

In milliseconds

ip_rate_limit: RateLimit

The default IP rate limit that will be imposed if Campaign.event_submission is None.

creators_whitelist: Vec<Address>

An optional whitelisted addresses for Campaign.creators.

If empty, any address will be allowed to create a Campaign.

validators_whitelist: Vec<ValidatorId>

An optional whitelisted Validator addresses for Campaign.validators.

If empty, any address will be allowed to be a validator in a Campaign.

admins: Vec<Address>chains: HashMap<String, ChainInfo>

The key of this map is a human-readable text of the Chain name for readability in the configuration file.

To get the chain of a token address use Config::find_chain_of().

To get a ChainInfo only by a ChainId use Config::find_chain().

NOTE: Make sure that a Token Address is unique across all Chains, otherwise Config::find_chain_of() will fetch only one of them and cause unexpected problems.

platform: PlatformConfiglimits: Limits
Implementations

impl Config

pub fn try_toml(toml: &str) -> Result<Self, TomlError>

Utility method that will deserialize a Toml file content into a Config.

Rather than relying on the toml crate directly, use this method instead.

pub fn find_chain(&self, chain_id: ChainId) -> Option<&ChainInfo>

Finds a Chain based on the ChainId.

pub fn find_chain_of(&self, token: Address) -> Option<ChainOf<()>>

Finds the pair of Chain & Token, given only a token Address.