Module primitives::constants::chain

Re-exports

pub use crate::AssetId;

pub use crate::Balance;

Constants

CORE_ASSET_ID

Core asset id

DISCOUNTED_FEE

Discounted XYK fee

MAXIMUM_BLOCK_WEIGHT

We allow for

MAX_IN_RATIO

Max fraction of pool to sell in single transaction

MAX_OUT_RATIO

Max fraction of pool to buy in single transaction

MIN_POOL_LIQUIDITY

Minimum pool liquidity

MIN_TRADING_LIMIT

Trading limit

WEIGHT_PER_SECOND

Type Definitions

Weight

Numeric range of a transaction weight.