primitives

Module div

Functions

cmux_public

Computes a CMUX: If c is 1, returns x_t, otherwise returns x_f.

cmux_public_many

div

div_assign_by_odd_public

Performs in-place division of a shared value by a public, odd value inside Z_{2^k}.

div_by_odd_public

Computes division by a public, odd value inside Z_{2^k}. The divisor must be invertible modulo 2^k, i.e. it has to be odd. The operation is implemented by multiplying with the modular inverse of the divisor.

div_multithreads

Division.

div_rem_const_public_arithmetic_i64

Division with i64 inputs and outputs. Note each party needs to do is different.

div_rem_const_public_arithmetic_many_i64

div_share_by_public

div_share_by_public_arithmetic_multithreads

fast version

non_restoring_division

Performs non-restoring division. N and D are inputs. bits is the number of bits for the quotient. Assumes T is large enough (at least 2*bits).

non_restoring_division_many

Batched version of non-restoring division.