primitives::is_default

Trait IsDefault

pub trait IsDefault {
    // Required method
    fn is_default(&self) -> bool;
}

This trait checks whether a variable equals to initialization value. For a variable equals to the initialization value, the world-state should treat is as None value.

Required Methods

fn is_default(&self) -> bool