Module primitives::report

App-specific crypto used for reporting equivocation/misbehavior in BABE and GRANDPA. Any rewards for misbehavior reporting will be paid out to this account.

Structs

ReporterAppCrypto

An AppCrypto type to allow submitting signed transactions using the reporting application key as signer.

Constants

KEY_TYPE

Key type for the reporting module. Used for reporting BABE and GRANDPA equivocations.

Type Definitions

ReporterId

Identity of the equivocation/misbehavior reporter.