Module primitives::sentry

Modules

campaign_create

campaign_list

campaign_modify

channel_list

message

units_for_slot

validator_messages

Structs

AccountingResponse

Channel Accounting response

AllSpendersQuery

AllSpendersResponse

Spender limits for all spenders on a Channel.

Analytics

AnalyticsResponse

Response returned when getting Analytics which returns the FetchedAnalytics.

ChannelPayRequest

Payouts to be performed for the given Channel.

DateHour

DateHour holds the date and hour (only).

DateHourError

FetchedAnalytics

GetLeafResponse

Get leaf response with the Merkle proof for the requested spender/earner.

InsertEventsRequest

Request body for posting new Events to a Campaign.

LastApproved

The last approved NewState and ApproveState accordingly to the validator.

LastApprovedQuery

GET /v5/channel/0xXXX.../last-approved query parameters

LastApprovedResponse

GET /v5/channel/0xXXX.../last-approved response

Pagination

SpenderResponse

Spender limits for a spender on a Channel.

SuccessResponse

UpdateAnalytics

ValidationErrorResponse

Enums

Event

All the Events available in the validator stack.

EventType

FetchedMetric

The value of the requested analytics crate::analytics::Metric.

Statics

CLICK

IMPRESSION