Trait primitives::signing::Sign

pub trait Sign: SecretKeyGen + PublicKeyGen {
    fn info(&self) -> SignInfo;

    fn sign(
        &self, 
        buf: &mut [u8], 
        data: &[u8], 
        secret_key: &[u8]
    ) -> Result<usize, Box<dyn Error + 'static>>;

    fn verify(
        &self, 
        data: &[u8], 
        sig: &[u8], 
        public_key: &[u8]
    ) -> Result<(), Box<dyn Error + 'static>>;
}

Required methods

fn info(&self) -> SignInfo

returns info about the signer

fn sign(
&self,
buf: &mut [u8],
data: &[u8],
secret_key: &[u8]
) -> Result<usize, Box<dyn Error + 'static>>

signs data in the buffer using the secret key and returns the signature length.

fn verify(
&self,
data: &[u8],
sig: &[u8],
public_key: &[u8]
) -> Result<(), Box<dyn Error + 'static>>

verify the signature for the data with the public key. Returns an error if the signature was invalid.

Loading content...