primitives::storage_key

Struct DeltaMptKeyPadding

pub struct DeltaMptKeyPadding(/* private fields */);

The padding is uniquely generated for DeltaMPT at each intermediate epoch, and itโs used to compute padding bytes for address and storage_key. The padding setup is against an attack where adversary artificially build deep paths in MPT.