Struct primitives::UnifiedNum

pub struct UnifiedNum(_);

Unified Number with a precision of 8 digits after the decimal point.

The number can be a maximum of u64::MAX (the underlying type), or in a UnifiedNum value 184_467_440_737.09551615. The actual number is handled as a unsigned number and only the display shows the decimal point.

This number is (de)serialized as a Javascript number which is f64. As far as the numbers don’t exceed 2**63, the Javascript number should be sufficient without losing precision

Examples

use primitives::UnifiedNum;
use serde_json::Value;

let unified_num = UnifiedNum::from(42_999_987_654_321);

// Printing the unified num will show the value and the decimal point with precision of `UnifiedNum::PRECISION` (i.e. `8`) numbers after the decimal point
assert_eq!("42999987654321", &unified_num.to_string());

assert_eq!("429999.87654321", &unified_num.to_float_string());

// Printing the Debug of unified num will show the value and the decimal point with precision of `UnifiedNum::PRECISION` (i.e. `8`) numbers after the decimal point
assert_eq!("UnifiedNum(429999.87654321)".to_string(), format!("{:?}", &unified_num));

// JSON Serializing and Deserializing the `UnifiedNum` yields a string without any decimal points
assert_eq!(Value::String("42999987654321".to_string()), serde_json::to_value(unified_num).unwrap());

Implementations

impl UnifiedNum

pub const PRECISION: u8 = 8u8

The precision of the UnifiedNum is 8 decimal numbers after the comma.

pub const MULTIPLIER: u64 = 100_000_000u64

The whole number multiplier when dealing with a UnifiedNum.

Examples

use primitives::UnifiedNum;

let whole_number = 8_u64; // we want to represent 8.00_000_000

assert_eq!(UnifiedNum::from_u64(800_000_000), UnifiedNum::from(whole_number * UnifiedNum::MULTIPLIER));

pub const DEBUG_DELIMITER: char = '.'

pub const ZERO: UnifiedNum = _

pub const ONE: UnifiedNum = _

The whole number 1 as a UnifiedNum. One (1) followed by exactly 8 zeroes (0).

1.00_000_000 100_000_000

pub fn div_floor(&self, other: &Self) -> Self

pub const fn from_u64(value: u64) -> Self

This method creates a UnifiedNum from an inner u64 value.

This method does not take into account precision of UnifiedNum!

pub const fn to_u64(self) -> u64

This method returns the inner u64 representation of the UnifiedNum.

This method does not take into account precision of UnifiedNum!

pub fn to_bignum(self) -> BigNum

pub fn checked_add(&self, rhs: &UnifiedNum) -> Option<Self>

pub fn checked_sub(&self, rhs: &UnifiedNum) -> Option<Self>

pub fn checked_mul(&self, rhs: &UnifiedNum) -> Option<Self>

pub fn checked_div(&self, rhs: &UnifiedNum) -> Option<Self>

pub fn checked_rem(&self, rhs: &UnifiedNum) -> Option<Self>

pub fn to_precision(self, precision: u8) -> BigNum

Transform the UnifiedNum precision 8 to a new precision

pub fn from_precision(amount: BigNum, precision: u8) -> Option<Self>

Transform the BigNum of a given precision to UnifiedNum with precision 8 If the resulting value is larger that what UnifiedNum can hold, it will return None

pub fn to_float_string(self) -> String