Trait primitives::traits::AMMHandlers

pub trait AMMHandlers<AccountId, AssetId, AssetPair, Balance> {
    fn on_create_pool(asset_pair: AssetPair);

    fn on_trade(
        amm_transfer: &AMMTransfer<AccountId, AssetId, AssetPair, Balance>, 
        liq_amount: Balance
    );

    fn on_trade_weight() -> Weight;
}

AMM handlers used by AMM pools to indicate various events.

Required methods

fn on_create_pool(asset_pair: AssetPair)

Register an asset to be handled by price-oracle pallet. If an asset is not registered, calling on_trade results in populating the price buffer in the price oracle pallet, but the entries are ignored and the average price for the asset is not calculated.

fn on_trade(
amm_transfer: &AMMTransfer<AccountId, AssetId, AssetPair, Balance>,
liq_amount: Balance
)

Include a trade in the average price calculation of the price-oracle pallet.

fn on_trade_weight() -> Weight

Known on_trade overhead. Needs to be specified here if we don’t want to make AMM pools tightly coupled with the price oracle pallet.