Enum primitives::transaction::TransactionError

pub enum TransactionError {

    AlreadyImported,
    ChainIdMismatch {
        expected: u32,
        got: u32,
        space: Space,
    },
    EpochHeightOutOfBound {
        block_height: u64,
        set: u64,
        transaction_epoch_bound: u64,
    },
    NotEnoughBaseGas {
        required: U256,
        got: U256,
    },
    Stale,
    TooCheapToReplace,
    LimitReached,
    InsufficientGasPrice {
        minimal: U256,
        got: U256,
    },
    InsufficientGas {
        minimal: U256,
        got: U256,
    },
    InsufficientBalance {
        balance: U256,
        cost: U256,
    },
    GasLimitExceeded {
        limit: U256,
        got: U256,
    },
    InvalidGasLimit(OutOfBounds<U256>),
    InvalidSignature(String),
    TooBig,
    InvalidRlp(String),
    ZeroGasPrice,
    FutureTransactionType,
    InvalidReceiver,
    TooLargeNonce,

}

Errors concerning transaction processing.

Variants

AlreadyImported

Transaction is already imported to the queue

ChainIdMismatch

Fields

expected: u32

got: u32

space: Space

Chain id in the transaction doesn’t match the chain id of the network.

EpochHeightOutOfBound

Fields

block_height: u64

set: u64

transaction_epoch_bound: u64

Epoch height out of bound.

NotEnoughBaseGas

Fields

required: U256

Absolute minimum gas required.

got: U256

Gas provided.

The gas paid for transaction is lower than base gas.

Stale

Transaction is not valid anymore (state already has higher nonce)

TooCheapToReplace

Transaction has too low fee (there is already a transaction with the same sender-nonce but higher gas price)

LimitReached

Transaction was not imported to the queue because limit has been reached.

InsufficientGasPrice

Fields

minimal: U256

Minimal expected gas price

got: U256

Transaction gas price

Transaction’s gas price is below threshold.

InsufficientGas

Fields

minimal: U256

Minimal expected gas

got: U256

Transaction gas

Transaction’s gas is below currently set minimal gas requirement.

InsufficientBalance

Fields

balance: U256

Senders balance

cost: U256

Transaction cost

Sender doesn’t have enough funds to pay for this transaction

GasLimitExceeded

Fields

limit: U256

Current gas limit

got: U256

Declared transaction gas

Transactions gas is higher then current gas limit

InvalidGasLimit(OutOfBounds<U256>)

Transaction’s gas limit (aka gas) is invalid.

InvalidSignature(String)

Signature error

TooBig

Transaction too big

InvalidRlp(String)

Invalid RLP encoding

ZeroGasPrice

FutureTransactionType

Transaction types have not been activated

InvalidReceiver

Receiver with invalid type bit.

TooLargeNonce

Transaction nonce exceeds local limit.