Module primitives::validator::messages

Validator Message Types

Structs

ApproveState

Generated by the follower when a NewState is generated by the leader and the state is signable and correct.

Heartbeat

Heartbeat sent to each Channel validator by the other validator of the Channel.

NewState

Generated by the Channel.leader on changed balances for the Channel.

RejectState

Generated by the Channel.follower on:

Enums

MessageError

MessageTypes

The message types used by validator.

Traits

Type