Crate primitives

Modules

asset

fee

traits

Structs

ExchangeIntention

Enums

IntentionType

Constants

CORE_ASSET_ID
Core asset id

MAX_IN_RATIO
Max fraction of pool to sell in single transaction

MAX_OUT_RATIO
Max fraction of pool to buy in single transaction

MIN_POOL_LIQUIDITY
Minimum pool liquidity

MIN_TRADING_LIMIT
Trading limit

Type Definitions

Amount
Signed version of Balance

AssetId
Type for storing the id of an asset.

Balance
Type for storing the balance of an account.

BlockNumber
An index to a block.

HighPrecisionBalance
Scaled Unsigned of Balance

LowPrecisionBalance

Moment
Type used for expressing timestamp.

Price
Price