Crate primitives

Modules

auth
A general purpose API for Cryptographic Primitives. This crate's aim is to provide an abstraction layer with extensible cryptographic primitives. Each primitive contains an info data structure for describing the constraints of the algorithm and at least one trait. Message Auth Code

cipher
Cipher

hash
Hash

key_derv_func
Key derive function

pbkdf
PBKDF

rng
Random Number Generator

signing
Asymmetric Signing