Crate primitives

Low-level types used throughout the core substrate code.

Structs

ParaId

Unique identifier of a parachain.

ValidationParams

Validation parameters for evaluating the parachain validity function.

Enums

XcmpMessageFormat

The aggregate XCMP message format.

Traits

DmpMessageHandler

Something that should be called when a downward message is received.

XcmpMessageHandler

Something that should be called for each batch of messages received over XCMP.

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

AuraId

An Aura authority identifier using S/R 25519 as its crypto.

Balance

Balance of an account.

Block

Block type.

BlockId

Block ID.

BlockNumber

An index to a block.

DigestItem

Digest item type.

Hash

A hash of some data used by the chain.

Header

Header type.

Index

Index of a transaction in the chain.

Moment

Type used for expressing timestamp.

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.

Timestamp

A timestamp: milliseconds since the unix epoch. u64 is enough to represent a duration of half a billion years, when the time scale is milliseconds.

UpwardMessage

A message from a parachain to its Relay Chain.