Module primitives::config

Modules

limits

std_duration_millis

Module for Config (de)serialization of std::time::Duration from and to milliseconds.

Structs

ChainInfo

Configured chain with tokens.

Config

Examples:

Limits

PlatformConfig

TokenInfo

Configured Token in a specific Chain. Precision can differ for the same token from one Chain to another.

TomlError

Errors that can occur when deserializing a type.

Enums

ConfigError

Environment

The environment in which the application is running Defaults to Environment::Development

Statics

GANACHE_CONFIG

Ganache (dev) configuration found in docs/config/ganache.toml

PRODUCTION_CONFIG

Production configuration found in docs/config/prod.toml

Functions

configuration

If no config_file path is provided it will load the Environment configuration. If config_file path is provided it will try to read and parse the file in Toml format.