Module primitives::postgres

Available on crate feature postgres only.

This module is available with the postgres feature.

Other places where you’d find mod postgres implementations is for many of the structs in the crate all of which implement tokio_postgres::types::FromSql, tokio_postgres::types::ToSql or From<&tokio_postgres::Row>

Statics

POSTGRES_CONFIG

Postgres configuration derived from the environment variables:

POSTGRES_DB

POSTGRES_DB environment variable - default: POSTGRES_USER

POSTGRES_HOST

POSTGRES_HOST environment variable - default: localhost

POSTGRES_PASSWORD

POSTGRES_PASSWORD environment variable - default: postgres

POSTGRES_POOL

A Postgres pool with reasonable settings:

POSTGRES_PORT

POSTGRES_PORT environment variable - default: 5432

POSTGRES_USER

POSTGRES_USER environment variable - default: postgres

Type Definitions

DbPool