Module primitives::sentry

Modules

campaign_create

campaign_list

campaign_modify

channel_list

message

Structs

AccountingResponse

Channel Accounting response A collection of all Accountings for a specific Channel

AllSpendersQuery

AllSpendersResponse

Analytics

AnalyticsResponse

Response returned when getting Analytics which returns the FetchedAnalytics.

ChannelPayRequest

DateHour

DateHour holds the date and hour (only).

DateHourError

FetchedAnalytics

InsertEventsRequest

Request body for posting new Events to a Campaign.

LastApproved

LastApprovedQuery

LastApprovedResponse

MessageResponse

Pagination

SpenderResponse

SuccessResponse

UpdateAnalytics

ValidationErrorResponse

ValidatorMessage

ValidatorMessagesCreateRequest

ValidatorMessagesListQuery

ValidatorMessagesListResponse

Enums

Event

EventType

FetchedMetric

The value of the requested analytics crate::analytics::Metric.

Statics

CLICK

IMPRESSION