Struct primitives::sentry::DateHour

pub struct DateHour<Tz: TimeZone> {
    pub date: Date<Tz>,
    pub hour: u32,
}

DateHour holds the date and hour (only).

It uses chrono::DateTime when serializing and deserializing. When serializing it always sets minutes and seconds to 0 (zero). When deserializing the minutes and seconds should always be set to 0 (zero), otherwise an error will be returned.

Fields
date: Date<Tz>hour: u32

hour is in the range of 0 - 23

Implementations

impl DateHour<Utc>

pub fn from_ymdh(year: i32, month: u32, day: u32, hour: u32) -> Self

Panics

When wrong inputs have been passed, i.e. for year, month, day or hour.

pub fn from_ymdh_opt(year: i32, month: u32, day: u32, hour: u32) -> Option<Self>

Makes a new DateHour from year, month, day and hour.

Returns None on invalid year, month, day or hour.

See chrono::NaiveDate::from_ymd_opt() & chrono::NaiveTime::from_hms_opt() for details

pub fn from_date_hour_opt(date: Date<Utc>, hour: u32) -> Option<Self>

pub fn now() -> Self

pub fn with_hour(&self, hour: u32) -> Option<Self>

pub fn with_day(&self, day: u32) -> Option<Self>

impl<Tz: TimeZone> DateHour<Tz>

pub fn to_datetime(&self) -> DateTime<Tz>

Creates a DateTime with minutes, seconds, nanoseconds set to 0 (zero)