primitives

Module state_root

Structs

StateRoot

The deferred state root consists of 3 parts: snapshot, delta_0, delta. when delta grows over threshold, snapshot and delta_0 is merged into new snapshot, and the delta becomes new delta_0.

Constants

MERKLE_NULL_NODE

The Merkle Hash for an empty MPT (either as a subtree or as a whole tree).

Type Aliases

MerkleHash